requirements.txt
django==4.2.*
djangorestframework==3.15.*
mysqlclient==2.2.*
redis==5.0.*
python-dotenv==1.0.*
torch==2.2.*
torchvision==0.17.*
torchxrayvision==1.0.*
onnx==1.16.*
onnxruntime==1.17.*
huggingface-hub==0.21.*
Pillow==10.*
numpy==1.26.*
pytest==8.*
pytest-django==4.*
